Feature lg

See merge request bdgroup/dmp-phase1!14

+++ Code/DMP-service/dmp/api/file.py
@@ -4,9 +4,10 @@
 # @Author  : SHTD
 import os
 
-from flask import Blueprint, jsonify, request, current_app
+import pandas as pd
+from flask import Blueprint,  request, current_app
 from dmp.utils import resp_hanlder, uuid_str
-from dmp.models import FromDownload
+from dmp.models import Forms
 
 file = Blueprint("file", __name__)
 
@@ -25,7 +26,7 @@ def upload(desc):
             filename = '%s%s' % (task, chunk)
             # 保存分片到本地
             upload_file.save(os.path.join(current_app.config.get("UPLOADED_PATH"), filename))
-            return resp_hanlder(result="chunk  %d OK"%chunk)
+            return resp_hanlder()
         except Exception as err:
             current_app.logger.error(err)
             return resp_hanlder(err=err)
@@ -35,11 +36,13 @@ def upload(desc):
 def success(desc):
     target_filename = request.args.get('filename')
     task = request.args.get('task_id')
+    filetype = request.args.get("filetype") or 1
     chunk = 0
     upload_path = current_app.config.get("UPLOADED_PATH")
     current_app.logger.info("%s%s%s" % (target_filename, task, upload_path))
+    tmp_filename = uuid_str() + "_TMP_" +target_filename
     finally_filename = uuid_str() + target_filename
-    with open(os.path.join(upload_path, finally_filename), 'wb') as target_file:
+    with open(os.path.join(upload_path, tmp_filename), 'wb') as target_file:
         while True:
             try:
                 filename = os.path.join(upload_path, '%s%d' % (task, chunk))
@@ -54,12 +57,19 @@ def success(desc):
             chunk += 1
             # 删除该分片，节约空间
             os.remove(filename)
+    if filetype == 1:
+        os.rename(os.path.join(upload_path, tmp_filename),os.path.join(upload_path, finally_filename))
+    elif filetype == 2:
+        data = pd.read_json(os.path.join(upload_path, tmp_filename))
+        data.to_csv(os.path.join(upload_path, finally_filename))
+    elif filetype == 3:
+        data = pd.read_excel(os.path.join(upload_path, tmp_filename))
+        data.to_csv(os.path.join(upload_path, finally_filename))
     current_app.logger.info(finally_filename)
     return resp_hanlder(result={"filename": finally_filename})
 
 
-class FormDownload(object):
-    pass
+
 
 
 @file.route("/dlcomplete/", methods=["GET"], defaults={"desc": {"interface_name": "文件下载完成","is_permission": False,"permission_belong": None}})
@@ -67,9 +77,16 @@ def dlcomplete(desc):
     if request.method == "GET":
         try:
             form_id = request.json.get("form_id")
-            form_ = FromDownload.get(form_id)
-            rm_filepath = form_.filepath
-            os.remove(rm_filepath)
+            form_ = Forms.get(form_id)
+            rm_filepath = form_.info_form.filepath
+            if os.path.exists(rm_filepath):
+                os.remove(rm_filepath)
+            dlform = form_.info_form
+            dlform.ftp_url = None
+            dlform.filepath = None
+            dlform.put()
             return resp_hanlder(result="OK")
+
+
         except Exception as err:
             return resp_hanlder(result="OK")
